requirements.txt
# local package
-e .

# external requirements
click~=7.1.1
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

numpy~=1.18.1
pandas~=1.0.2
pathlib~=1.0.1
opencv-python~=4.2.0.32
pillow~=2.6.0
tqdm~=4.42.1
torch~=1.4.0
matplotlib~=3.1.3
scikit-learn~=0.22.1
html4vision~=0.4.0
tensorflow~=2.0.0
scipy~=1.4.1
soundfile~=0.10.3.post1
pyaudio~=0.2.11
torchvision~=0.5.0
tensorboardx~=2.0
imutils~=0.5.3
seaborn~=0.10.0
swifter~=0.302
moviepy~=1.0.2
wget~=3.2
mtcnn~=0.1.0
setuptools~=46.0.0
dominate~=2.5.1
flask~=1.1.1
pytest~=5.4.1
pyyaml~=5.3.1
typing_extensions~=3.7.4.1
librosa~=0.7.2
dask~=2.13.0
sortedcontainers~=2.1.0
keras~=2.3.1